PROACTIVE-571: Forgot to check in GCMVirtualNodERemoteObjectAdapter in the previous commit. Timeout is now a long instead of an int.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@10996 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Extensions/org/objectweb/proactive/extensions/gcmdeployment/core/GCMVirtualNodeRemoteObjectAdapter.java
@@ -160,7 +160,7 @@ public class GCMVirtualNodeRemoteObjectAdapter extends Adapter<GCMVirtualNode> i
         vn.waitReady();
     }
 
-    public void waitReady(int timeout) throws ProActiveTimeoutException {
+    public void waitReady(long timeout) throws ProActiveTimeoutException {
         vn.waitReady(timeout);
     }
